添加HDL_EXTI.C文件，实现三个正转目标值的设定

+++ BSL/BSL_ENCODER.h
@@ -4,7 +4,7 @@
 typedef struct
 { 
 
-	uint32_t distance;
+	float distance;
 	uint8_t speed;
 	uint8_t quanshu;
 }ENCODER_NUM;
